+++ aksel.nav.no/website/components/sanity-modules/code-examples/CodeSandbox.tsx
@@ -1,4 +1,4 @@
-import { Edit } from "@navikt/ds-icons";
+import { PencilIcon } from "@navikt/aksel-icons";
 import { getParameters } from "codesandbox/lib/api/define";
 
 const indexTsx = `import React from 'react';
@@ -29,6 +29,7 @@ export const CodeSandbox = ({ code }: { code: string }) => {
             "@navikt/ds-react-internal": "latest",
             "@navikt/ds-css-internal": "latest",
             "@navikt/ds-icons": "latest",
+            "@navikt/aksel-icons": "latest",
           },
         } as any,
         isBinary: false,
@@ -58,9 +59,9 @@ export const CodeSandbox = ({ code }: { code: string }) => {
       <input type="hidden" name="query" value="module=App.js" />
       <button
         type="submit"
-        className="flex items-center justify-center gap-2 rounded-sm px-3 py-2 hover:bg-gray-100 focus:outline-none focus-visible:shadow-focus-inset active:bg-gray-200"
+        className="focus-visible:shadow-focus-inset flex items-center justify-center gap-1 rounded-sm px-3 py-2 hover:bg-gray-100 focus:outline-none active:bg-gray-200"
       >
-        <Edit aria-hidden />
+        <PencilIcon aria-hidden fontSize="1.5rem" />
         CodeSandbox
       </button>
     </form>